requirements.txt
streamlit==1.41.0
tensorflow==2.18.0
numpy==1.26.4
Pillow==11.0.0
